requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
fontawesomefree==6.5.1
phonenumbers==8.13.27
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
wheel==0.41.3
